LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/fetch.https.sub.any.worker-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/echo-as-json.py

PASS Same-origin fetch
FAIL Same-site fetch promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Cross-site fetch promise_test: Unhandled rejection with value: object "TypeError: URL is not valid or contains user credentials."
PASS Same-origin mode
PASS CORS mode
PASS no-CORS mode
PASS Same-origin fetch: sec-fetch-dest
PASS Same-origin fetch: sec-fetch-mode
FAIL Same-origin fetch: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-origin fetch: sec-fetch-user
PASS Same-origin mode: sec-fetch-dest
PASS Same-origin mode: sec-fetch-mode
FAIL Same-origin mode: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-origin mode: sec-fetch-user
PASS CORS mode: sec-fetch-dest
PASS CORS mode: sec-fetch-mode
FAIL CORS mode: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS CORS mode: sec-fetch-user
PASS no-CORS mode: sec-fetch-dest
PASS no-CORS mode: sec-fetch-mode
FAIL no-CORS mode: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS no-CORS mode: sec-fetch-user
